Fixes for cq.cfg
Remove the extra cq_name and add gerrit {} which is required.

A fun bug happened because our CQ config landed both global/ and
branch/ in the same commit. LUCI tries to discover where cq.cfg with
refs.cfg, however the processing of cq.cfg is done before refs.cfg. This
means that when everything was landed at once, the processing of cq.cfg
didn't where to look, and then refs.cfg was processed, but too late.

Change-Id: I9024278c6a32c806f4ed25b59f54a0365ac24566